Move Benchmarks class to a file
